--- services/web/server/tests/unit/isolated/test_projects_models.py
@@ -1,53 +1,0 @@
-# pylint:disable=unused-variable
-# pylint:disable=unused-argument
-# pylint:disable=redefined-outer-name
-
-import datetime
-
-import pytest
-from simcore_service_webserver.projects._db_utils import (
-    convert_to_db_names,
-    convert_to_schema_names,
-)
-
-
-@pytest.fixture
-def fake_schema_dict():
-    return {
-        "anEntryThatUsesCamelCase": "I'm the entry",
-        "anotherEntryThatUsesCamelCase": "I'm also an entry",
-    }
-
-
-@pytest.fixture
-def fake_db_dict():
-    return {
-        "an_entry_that_uses_snake_case": "I'm the entry",
-        "another_entry_that_uses_snake_case": "I'm also an entry",
-    }
-
-
-def test_convert_to_db_names(fake_schema_dict):
-    db_entries = convert_to_db_names(fake_schema_dict)
-    assert "an_entry_that_uses_camel_case" in db_entries
-    assert "another_entry_that_uses_camel_case" in db_entries
-
-
-def test_convert_to_schema_names(fake_db_dict):
-    fake_email = "fakey.justafake@fake.faketory"
-    db_entries = convert_to_schema_names(fake_db_dict, fake_email)
-    assert "anEntryThatUsesSnakeCase" in db_entries
-    assert "anotherEntryThatUsesSnakeCase" in db_entries
-    # test date time conversion
-    date = datetime.datetime.utcnow()
-    fake_db_dict["time_entry"] = date
-    db_entries = convert_to_schema_names(fake_db_dict, fake_email)
-    assert "timeEntry" in db_entries
-    assert db_entries["timeEntry"] == "{}Z".format(
-        date.isoformat(timespec="milliseconds")
-    )
-    # test conversion of prj owner int to string
-    fake_db_dict["prj_owner"] = 1
-    db_entries = convert_to_schema_names(fake_db_dict, fake_email)
-    assert "prjOwner" in db_entries
-    assert db_entries["prjOwner"] == fake_email
